requirements.txt
pexels-api
ffmpeg-python
Pillow
moviepy
requests
